requirements.txt
# This file is @generated by PDM.
# Please do not edit it manually.

annotated-types==0.7.0
asgiref==3.8.1
attrs==25.3.0
brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1; platform_python_implementation != "PyPy"
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.1
cloudinary==1.44.0
colorama==0.4.6; sys_platform == "win32" or platform_system == "Windows"
cryptography==44.0.3
distlib==0.3.9
dj-database-url==2.3.0
django==5.2.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-extensions==4.1
django-filter==25.1
djangorestframework==3.16.0
drf-spectacular==0.28.0
filelock==3.18.0
gunicorn==23.0.0
h11==0.16.0
identify==2.6.12
idna==3.10
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
mysqlclient==2.2.7
nodeenv==1.9.1
packaging==25.0
passage-identity==3.0.1
pillow==11.2.1
platformdirs==4.3.8
pre-commit==4.2.0
psycopg2-binary==2.9.10
pycparser==2.22; platform_python_implementation != "PyPy"
pydantic==2.11.5
pydantic-core==2.33.2
pydotplus==2.0.2
pyjwt==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-magic==0.4.27; sys_platform == "linux"
python-magic-bin==0.4.14; sys_platform == "win32"
pyyaml==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
typing-extensions==4.13.2
typing-inspection==0.4.1
tzdata==2025.2; sys_platform == "win32"
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
virtualenv==20.31.2
whitenoise[brotli]==6.9.0
